requirements.txt
google-genai
PyPDF2
pdfplumber
streamlit
python-dotenv
